requirements.txt
APScheduler==3.6.3
certifi==2020.11.8
chardet==3.0.4
emoji==0.6.0
Flask==1.1.2
geopy==2.0.0
idna==2.10
jsonconfig-tool==1.5
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.3
requests==2.25.0
urllib3==1.26.2
